requirements.txt
celery==4.4.6
django-celery-beat==2.2.0
django-configurations==2.2
django-cors-headers==3.7.0
django-extensions==2.0.6
django-redis==4.12.1
django-utils-six==2.0
django==3.1.3
djangorestframework==3.12.2
google-api-python-client==2.22.0
ipython==7.28.0
oauth2client==4.1.3
psycopg2-binary==2.8.6